gpu: gmc: introduce GMC framework
GMC is a shortcut for 'Graphic Memory Compression'. This is a name of a simple
framework developed especially for GPU kernel drivers which implement some
variant of GPU memory reclaiming techniques.

GMC provides an infrastructure for GPU drivers and allow to register callbacks
exported to user space as a debugfs interface. It also implements a concept of
compressed page storage similar to ZRAM but without an overhead of the block
layer.

GMC creates the following strcuture in the file system:

	/sys/kernel/debug/gmc/
			|
			+- device0/
			|	|
			|	+- compress
			|	+- decompress
			|	+- storage_stat
			|
			+- device1/
			|	|
			|	+- compress
			|	+- decompress
			|	+- storage_stat
			:
			|
			+- deviceN/
				|
				+- compress
				+- decompress
				+- storage_stat

The files 'compress' and 'decompress' are used to force a GPU driver associated
with a particular 'gmc device' to reclaim and bring back (compress and
decompress) graphic memory associated with some process. TGID of the process
should be written to one of the files.

Reading of the file 'storage_stat' allows to monitor internal statistics of
compressed storage associated with the device.

For more information on API read the files 'include/linux/gmc.h' and
'include/linux/gmc_storage.h'.

Change-Id: I691e8a668b7b5b5f3a4a7b113ce02d084bfe1ace
Signed-off-by: Sergei Rogachev <s.rogachev@samsung.com>
Signed-off-by: Alexander Yashchenko <a.yashchenko@samsung.com>