gpu/mali: change vma rw_spinlock to rw_sem
During operation of the MALI_MEM_OS reclaimer on compression and decompression
of graphical memory regions this is necessary to hold a lock protecting a tree
of MALI virtual memory areas. At the same time there are a number of other
primitives like mutexes that must be held too (and these are the sleeping
primitives), moreover allocation of a page on the decompression can sleep.

It is well known that it is forbidden* to sleep in an atomic context and in a
spinlock-protected critical section, so the critical section context should be
made non-atomic by replacement of the rw-spinlock to rw-semaphore.

[*] Spinlocks make an atomic context by incrementing the preemption counter, so
preemption of the task after execution of an interrupt handler is impossible,
but it is still possible to call schedule() directly, say from some sleeping
primitive. In such case the task goes to the sleeping state and cannot get the
control back (be woken up on the same CPU), because the preemption counter is
not zero. There are other examples of live-locking and dead-locking when someone
sleeps in the atomic context.

NOTE:
Unfortunately this change should follow the coding style specific to the files
emerged in the revision R5P2 with lots of "\r\n" ends of lines, etc.

Change-Id: Ie3dafcc15e74bc1e844ca0a0ac131a972142cdf1
Signed-off-by: Sergei Rogachev <s.rogachev@samsung.com>