mmc: sdhost: fix scheduling while atomic
The function mmc_regulator_set_ocr() cannot be called under the spinlock,
because internally it calls regulator_disable() which uses a sleeping primitive:
regulator_dev->mutex.

The patch unlocks the spinlock sdhost_host->lock to avoid scheduling in atomic
context and prevent possible consequent live-locks.

It is done totally the same way as it is done in drivers/mmc/host/sdhci.c and
many other places.

Change-Id: I2a7b893124efb2a515a3d55706d9a292a3d27edd
Signed-off-by: Sergei Rogachev <s.rogachev@samsung.com>