gpu/mali: introduce necessary GPU VMA functions
Implement a function to walk through the tree of GPU's virtual memory area
structures associated with some session. Provide a possibility to pass a
callback to the function.

Introduce a necessary non-interlocked functions.

NOTE:
Unfortunately this change should follow the coding style specific to the files
emerged in the revision R5P2 with lots of "\r\n" ends of lines, etc.

Change-Id: I0b58af2423d76afd340f029dffbf10250b299be9
Signed-off-by: Sergei Rogachev <s.rogachev@samsung.com>