gpu/mali: introduce PF-toggling function
Implement a function to toggle page faults on a specific GPU MMU (i.e. group)
by manipulation on read and write permission bits of ATEs.

Be careful using this function because in opposite to the myth that this isn't
needed to drop the TLB cache after manipultion on such bits (when a bus address
is not changed), such TLB invalidation is really necessary.

[ initial development ]
Signed-off-by: Krzysztof Kozlowski <k.kozlowski@samsung.com>
[ minor changes ]
Signed-off-by: Sergei Rogachev <s.rogachev@samsung.com>

Change-Id: I081efcf334d478d76d64cec85a3d6d3bc718e97c