gpu/mali: introduce PF state checking function
Introduce a function to check that a GPU group is in a page fault state. Such
function is needed to determine cases when this is safe to manipulate the TLB
cache without going to the MMU stall mode on some specific group. See one of the
following patches.

The function is also helpful to determine that nobody else has already handled a
page fault on the same group for some reason.

[ initial development ]
Signed-off-by: Krzysztof Kozlowski <k.kozlowski@samsung.com>
[ minor changes ]
Signed-off-by: Sergei Rogachev <s.rogachev@samsung.com>

Change-Id: If506cdfb1d96a66615617ad75dcccda5697726ec