gpu/mali: fix zapping TLB after allocation unmap
By default if TLB zapping was done unsuccessfully on freeing a memory
allocation this meant an error and the corresponding job was killed.

Zapping can be unsuccessful only if the core MALI code cannot stall an MMU. An
MMU cannot be stalled in two cases:

    1. More likely: MMU is in the page fault state;
    2. Stall mode timeout due to high load on the bus.

After introduction of valid page fault handling this is O.K. when an MMU cannot
be stalled because it is in the page fault state. Thus valid and invalid states
can be described by the following table:

  Validity             | V | V | I | X |
-----------------------+---+---+---+---+---
  In page fault        | x |   |   | x |
-----------------------+---+---+---+---+---
  Stalled successfully |   | x |   | x |
-----------------------+---+---+---+---+---
                       |   |   |   |   |

V: valid, I: invalid, X: impossible.

Thus, when a page fault is handled this is safe to do MMU operations without MMU
stall mode and the function mali_group_zap_session() can return MALI_TRUE.

Change-Id: Ic1fbfbb83793e63d5ce78d768e489d1589e11248
Signed-off-by: Sergei Rogachev <s.rogachev@samsung.com>