gpu/mali: change sessions list locking policy
By default a list of sessions in the Utgard driver was protected by a spin lock.
Because implementation of a reclaimer for MALI_MEM_OS allocations requires
sleeping under this lock it has been decided to replace the spin lock with a
read-write semaphore (rw_mutex in terms of the Mali driver).

Change-Id: Icd4076b5f811a3ea38aea900a92d6936dfaaef6c
Signed-off-by: Sergei Rogachev <s.rogachev@samsung.com>