gpu/mali: introduce os memory reclaimer
This commit implements a memory reclaim facility for buffers allocated by a
mem_os allocator. Such buffers typically contain inputs for GPU jobs (textures,
various scene-related data structures, etc).

Compression is triggered from the user space by writing a process ID to the
virtual file /sys/kernel/device/gmc/deviceX/compress, where X is a number of a
GMC device.

The implementation uses the GMC subsystem for organization of kernel/user space
interface and storage of compressed pages. Amount of available GMC devices
depends on a number of GPU drivers/devices active in the system. Typically this
quantity is 1, so all the manipulations are done on the "device0" node.

In Tizen 3.0 it is planned to use a modified ResourceD user space daemon to
watch the processes that go to background, so their graphical memory can be
compressed without significant performance degradation. This policy is an
opposite to LRU when detecting a working set is done statistically. In this case
the solution knows exactly which memory regions are temporarily unused. Such
fact neglects a performance overhead in most cases.

Decompression is done on demand and triggered by CPU or GPU page fault
exceptions.

Typical MALI workflow follows a number of simple steps:

	1. A user space library (libmali.so) requests an allocation of
	MALI_MEM_OS region for some graphical application via IOCTL call to the
	driver;
	2. The driver allocates a region and maps the underlying pages to GPUs
	page directory corresponding to the GPU session of the requesting
	process;
	3. The user space library maps the region using the mmap() system call
	for access from the CPU side;
	4. It places some data (textures, shaders, structures, etc) to the
	memory region;
	5. The library requests execution of a set of GPU jobs via IOCTL;
	6. Jobs are executed on a set of GPU cores (groups);
	7. The result is reported to the calling side.

At any point in time from descibed above ResourceD can request the kernel side
to compress graphical memory of an application. To prevent accesses from CPU and
GPU side physical memory is unmapped from its virtual views. Thus, on any access
from both devices (CPU and GPU) a kind of a page fault exception is triggered
and a corresponding fault handler is called. This means: there is a guarantee
that data from the pages underlying the allocations can be transparently
compressed without any negative effect to the users of these pages.

Most of the changes implementing the functionality itself are located in the
file mali_memory_os_reclaim.c, a number of functions are spread around in the
executor code (mali_executor.c), memory manager (mali_memory_manager.c) and
mem_os allocator (mali_memory_os_alloc.c).

To enable the functionality the option CONFIG_MALI_MEM_OS_RECLAIM should be
turned on. This option depends on CONFIG_GPU_GMC_GENERIC that should be enabled
too.

Signed-off-by: Sergei Rogachev <s.rogachev@samsung.com>
Change-Id: I07f94747acf226c55571750c6c03ff92155d9236